move out embedded rules from checkers.go
